fix: deposits & parent balance (#193)
## Summary

- The `account deposit` method no longer worked, resulting in an error
for `"NotRegisteredSubnet()"`. This is because we were passing the
`EVMSubnet` for the parent, which has no context for which subnet ID the
deposit is happening for. That is, it's subnet ID was _only_ the root
and no children, which is why we saw this error.
- The `account info` method logs the native coin balance on the parent.
This is confusing because it should log the erc20 supply source
balance—since native tFIL (or anvil ETH) has no meaning wrt the account.

## Details

If you run `hoku account deposit 1`, it works. Also, if you try to get
your balance, it now shows the erc20 balance—e.g., on localnet, this
used to show the anvil ETH balance with a value of `9999.99`:
```json
{
  "address": "0x90f79bf6eb2c4f870365e785982e1f101e93b906",
  "fvm_address": "t410fsd3zx5xlfrhyoa3f46czqlq7capjhoighmzagaq",
  "sequence": 2,
  "balance": "5003.0",
  "parent_balance": "97.0"
}
```

Note: this fixes things as-is, but there might be better ways to handle
it. E.g., we call `Account::supply_source_balance` for the parent
subnet, and this is okay since we only have one parent. But, with
hierarchical subnets, we'd probably have to handle this a bit
differently and check what the configured supply source is.