feat: show eth txs (#192)
Unifies the transaction response type for SDK methods to Ethereum style.

This is a bit of a trick to use the same CometBFT -> Ethereum tx
conversions from the IPC Ethereum API to return the Ethereum tx
equivalent for normal CometBFT txs.

This enables users to lookup tx hashes on the Blockscout (EVM-based)
chain explorer.

Depends on https://github.com/hokunet/ipc/pull/469.

Signed-off-by: Sander Pick <sanderpick@gmail.com>