feat: allow direct setting of config admin (#206)
@islamaliev you can use this to set the config admin. note: you can
already set the admin by just setting the config (the signer will become
admin if it's not already set). but we should add this command as well.

@sam701 , this reminds me... at some point during / after setting up a
new testnet... we should make sure to set the admin with this command.
ideally admin is a new private key that's only used for this purpose. on
mainnet, we'll make it a multisig.

Signed-off-by: Sander Pick <sanderpick@gmail.com>