fix: update token credit rate type (#228)
There was a fix to the Token Credit Rate type in
https://github.com/recallnet/ipc/pull/561/.
This PR makes the SDK and CLI compatible with those changes to `ipc`.  
Note: this change will **not** work with the testnet until after the
next reset. This means after this change is merged, we will need to
ensure people using the testnet know they need to be using the `v0.1.4`
tag, not the main branch.