feat: add dynamic network config (#233)
This PR adds an additional CLI flag `--network-config-file` that takes a
path to a TOML file that specifies a network. The default config file is
`~/.config/recall/networks.toml` that is automatically created if not exists.

The PR also deprecates `Localnet` `NetworkConfig`, because each network
will have different config.

Close #231