feat: ci using localnet image (#237)
This PR adds plumbing to setup a `localnet` Docker container and run
tests against it. A single sample test using the Recall CLI is included.

To be able to integrate with the SDK, we'll have to update the SDK (in a
separate PR) to be able to use configuration from the environment.

The Dagger plumbing allows many other CI combinations to be integrated
(e.g. bash scripts, or other test containers, etc.)

Closes #183.

---------

Signed-off-by: Mohsin Zaidi <2236875+smrz2001@users.noreply.github.com>